+++ indexer/runes/table/runes_status.go
@@ -9,7 +9,7 @@ import (
 type RunesStatus struct {
 	Table[pb.RunesStatus]
 	Version       string
-	Height        uint64
+	Height        int
 	Number        uint64
 	ReservedRunes uint64
 }
@@ -31,7 +31,7 @@ func (s *RunesStatus) Init() (ret bool) {
 			common.Log.Panicf("RunesStatus.Init-> db version inconsistent %s", store.DB_VERSION)
 		}
 		s.Version = stats.Version
-		s.Height = stats.Height
+		s.Height = int(stats.Height)
 		s.Number = stats.Number
 		s.ReservedRunes = stats.ReservedRunes
 		ret = true
@@ -43,7 +43,7 @@ func (s *RunesStatus) UpdateDb() {
 	key := []byte(store.STATUS_KEY)
 	value := &pb.RunesStatus{
 		Version:       s.Version,
-		Height:        s.Height,
+		Height:        uint64(s.Height),
 		Number:        s.Number,
 		ReservedRunes: s.ReservedRunes,
 	}
@@ -54,7 +54,7 @@ func (s *RunesStatus) Update() {
 	key := []byte(store.STATUS_KEY)
 	value := &pb.RunesStatus{
 		Version:       s.Version,
-		Height:        s.Height,
+		Height:        uint64(s.Height),
 		Number:        s.Number,
 		ReservedRunes: s.ReservedRunes,
 	}
